refactor: 유저 리스트 기능 클래스로 분리
- FileHandler에서 분기문으로 처리하지 않고 클래스로 분리하였다.
