+++ src/main/java/http/handler/CommandMatcher.java
@@ -7,7 +7,8 @@ public class CommandMatcher {
     public enum Command {
         USER_CREATE("/user/create", new UserCreateHandler()),
         USER_LOGIN("/user/login", new UserLoginHandler()),
-        USER_LOGOUT("/user/logout", new UserLogoutHandler());
+        USER_LOGOUT("/user/logout", new UserLogoutHandler()),
+        USER_LIST("/user/list", new UserListHandler());
 
         public final String filePath;
         public final CommandHandler commandHandler;
