+++ src/main/java/http/handler/UserCreateHandler.java
@@ -20,34 +20,4 @@ public class UserCreateHandler extends CommandHandler {
         UserDatabase.addUser(user);
         responseManager.setRedirectResponse(BASE_PATH);
     }
-
-    @Override
-    public void handlePutRequest() {
-        responseManager.setErrorResponse(Status.BAD_REQUEST);
-    }
-
-    @Override
-    public void handleDeleteRequest() {
-        responseManager.setErrorResponse(Status.BAD_REQUEST);
-    }
-
-    @Override
-    public void handleHeadRequest() {
-        responseManager.setErrorResponse(Status.BAD_REQUEST);
-    }
-
-    @Override
-    public void handleConnectRequest() {
-        responseManager.setErrorResponse(Status.BAD_REQUEST);
-    }
-
-    @Override
-    public void handleTraceRequest() {
-        responseManager.setErrorResponse(Status.BAD_REQUEST);
-    }
-
-    @Override
-    public void handlePatchRequest() {
-        responseManager.setErrorResponse(Status.BAD_REQUEST);
-    }
 }
